+++ jfrog/artifactory.py
@@ -17,7 +17,7 @@ logging.basicConfig(
 
 def artifactory_ping(url, token):
     """
-    This function is intented to get the health info of Jfrog Platform
+    This function is intented to get the health info of JFrog Platform
 
     Parameters
     ----------
@@ -32,7 +32,8 @@ def artifactory_ping(url, token):
         reponse
     """
     HEADERS.update({"Authorization": "Bearer " + token})
-    url = utilities.__validate_url(url)  # pylint: disable=W0212
+    url = utilities.__validate_url(  # pylint: disable=W0212:protected-access
+        url)
     urltopost = url + "/artifactory/api/system/ping"
     try:
         response = requests.get(urltopost, headers=HEADERS, timeout=30)
@@ -49,12 +50,12 @@ def artifactory_ping(url, token):
 
 def artifactory_version(url, token):
     """
-    This function is intented to get the version info of Jfrog Platform
+    This function is intented to get the version info of JFrog Platform
 
     Parameters
     ----------
     arg1 : str
-        base URL of Jfrog PLatform
+        base URL of JFrog Platform
     arg2 : str
         access or identity token of admin account
 
@@ -64,14 +65,15 @@ def artifactory_version(url, token):
         version of artifactory
     """
     HEADERS.update({"Authorization": "Bearer " + token})
-    url = utilities.__validate_url(url)  # pylint: disable=W0212
+    url = utilities.__validate_url(  # pylint: disable=W0212:protected-access
+        url)
     urltopost = url + "/artifactory/api/system/version"
     response = requests.get(urltopost, headers=HEADERS, timeout=30)
     if response.ok:
         versioninfo = response.json()
-        logging.info(
-            "Your Artifactory Instance is currently running %s", versioninfo['version'])
         version = versioninfo['version']
+        logging.info(
+            "Your Artifactory Instance is currently running %s", version)
     else:
         logging.error("Could not determin the Artifactory version")
         print(tabulate(response.json()))
@@ -81,12 +83,12 @@ def artifactory_version(url, token):
 
 def get_license_details(url, token):
     """
-    This function is intented to get the license info of Jfrog Platform
+    This function is intented to get the license info of JFrog Platform
 
     Parameters
     ----------
     arg1 : str
-        base URL of Jfrog Platform
+        base URL of JFrog Platform
     arg2 : str
         access or identity token of admin account
 
@@ -96,7 +98,8 @@ def get_license_details(url, token):
         dictionary of license information
     """
     HEADERS.update({"Authorization": "Bearer " + token})
-    url = utilities.__validate_url(url)  # pylint: disable=W0212
+    url = utilities.__validate_url(  # pylint: disable=W0212:protected-access
+        url)
     urltopost = url + "/artifactory/api/system/license"
     response = requests.get(urltopost, headers=HEADERS, timeout=30)
     if response.ok:
@@ -110,12 +113,12 @@ def get_license_details(url, token):
 
 def get_ha_nodes(url, token):
     """
-    This function is intented to get the count of nodes in a Jfrog Platform HA setup
+    This function is intented to get the count of nodes in a JFrog Platform HA setup
 
     Parameters
     ----------
     arg1 : str
-        base URL of Jfrog Platform
+        base URL of JFrog Platform
     arg2 : str
         access or identity token of admin account
 
@@ -125,7 +128,8 @@ def get_ha_nodes(url, token):
         number of nodes
     """
     HEADERS.update({"Authorization": "Bearer " + token})
-    url = utilities.__validate_url(url)  # pylint: disable=W0212
+    url = utilities.__validate_url(  # pylint: disable=W0212:protected-access
+        url)
     urltopost = url + "/artifactory/api/system/licenses"
     response = requests.get(urltopost, headers=HEADERS, timeout=30)
     if response.ok:
@@ -145,7 +149,7 @@ def get_repo_count(url, token, repository_type):
     Parameters
     ----------
     arg1 : str
-        base URL of Jfrog PLatform
+        base URL of JFrog Platform
     arg2 : str
         access or identity token of admin account
     arg3 : str
@@ -158,7 +162,8 @@ def get_repo_count(url, token, repository_type):
         number of repositories
     """
     HEADERS.update({"Authorization": "Bearer " + token})
-    url = utilities.__validate_url(url)  # pylint: disable=W0212
+    url = utilities.__validate_url(  # pylint: disable=W0212:protected-access
+        url)
     urltopost = url + f'/artifactory/api/repositories?type = {repository_type}'
     response = requests.get(urltopost, headers=HEADERS, timeout=30)
     if response.ok:
@@ -179,7 +184,7 @@ def get_storage_info(url, token):
     Parameters
     ----------
     arg1 : str
-        base URL of Jfrog PLatform
+        base URL of JFrog Platform
     arg2 : str
         access or identity token of admin account
 
@@ -189,7 +194,8 @@ def get_storage_info(url, token):
         dictionary of storage information
     """
     HEADERS.update({"Authorization": "Bearer " + token})
-    url = utilities.__validate_url(url)  # pylint: disable=W0212
+    url = utilities.__validate_url(  # pylint: disable=W0212:protected-access
+        url)
     urltopost = url + '/artifactory/api/storageinfo'
     response = requests.get(urltopost, headers=HEADERS, timeout=30)
     if response.ok:
@@ -200,3 +206,73 @@ def get_storage_info(url, token):
         storageinfo = {'binariesCount': '0', 'binariesSize': '0 GB', 'artifactsSize': '0 GB',
                        'optimization': '0%', 'itemsCount': '0', 'artifactsCount': '0'}
     return storageinfo
+
+
+def rename_repo(url, token, old_repo_name, new_repo_name, ptype, action='copy', delete=False):
+    """
+    This function will rename a repository by creating a new repo and moving the contents
+    New repo will be created with default values
+
+    Parameters
+    ----------
+    arg1 : str
+        base URL of JFrog Platform
+    arg2 : str
+        access or identity token of admin account
+    arg3 : str
+        old repo name
+    arg4 : str
+        new repo name
+    arg5 : str
+        package type for the new repo
+    arg6 : str
+        action to be taken with the contents
+        valid options are copy or move
+    arg7 : bool
+        True or False flag to delete the old repo when action is complete
+
+    """
+    HEADERS.update({"Authorization": "Bearer " + token})
+    url = utilities.__validate_url(  # pylint: disable=W0212:protected-access
+        url)
+    if action != 'action' or action != 'move':
+        action = 'copy'
+    # get type of old repo to create new repo as same type
+    urltopost = url + f'/artifactory/api/repositories/{old_repo_name}'
+    response = requests.get(urltopost, headers=HEADERS, timeout=30)
+    json_object = response.json()
+    rtype = json_object["rclass"]
+    # create new repo with default settings
+    urltopost = url + f'/artifactory/api/repositories/{new_repo_name}'
+    data = utilities.__set_repo_data(new_repo_name, utilities.__setlayout(  # pylint: disable=W0212:protected-access
+        ptype), ptype, rtype)
+    response = requests.put(urltopost, headers=HEADERS, data=data, timeout=30)
+    if response.ok:
+        logging.info(response.text)
+        # copy or move contents from old repo to new repo
+        urltopost = url + \
+            f'/artifactory/api/storage/{old_repo_name}?list&deep=1'
+        response = requests.get(urltopost, headers=HEADERS, timeout=30)
+        try:
+            json_object = response.json()
+            items = json_object["files"]
+            for item in items:
+                uri = item.get('uri')
+                urltopost = url + \
+                    f'/artifactory/api/{action}/{old_repo_name}{uri}?to=/{new_repo_name}{uri}'
+                response = requests.post(
+                    urltopost, headers=HEADERS, timeout=30)
+                logging.info(utilities.__get_msg(response, 'messages')  # pylint: disable=W0212:protected-access
+                             )
+            # delete old repo if flag set to True
+            if delete:
+                urltopost = url + \
+                    f'/artifactory/api/repositories/{old_repo_name}'
+                response = requests.delete(
+                    urltopost, headers=HEADERS, timeout=30)
+                logging.info(response.text)
+        except KeyError:
+            logging.error('%s not found', old_repo_name)
+    else:
+        logging.error(utilities.__get_msg(response, 'errors')  # pylint: disable=W0212:protected-access
+                      )